switch endpoints
