+++ lib/domain/dtos/EndRunDto.js
@@ -35,7 +35,7 @@ const QueryDto = Joi.object({
 });
 
 const ParamsDto = Joi.object({
-    runId: Joi.string(),
+    runNumber: Joi.string(),
 });
 
 const EndRunDto = Joi.object({
