+++ lib/public/views/Runs/Details/RunDetailsModel.js
@@ -140,9 +140,9 @@ export class RunDetailsModel extends Observable {
         this._runDetails = RemoteData.loading();
         this.notify();
         if (this._runNumber) {
-            return this._handleFetchRemoteRun(() => getRemoteData(`/api/v2/runs/${this._runNumber}`));
+            return this._handleFetchRemoteRun(() => getRemoteData(`/api/runs/${this._runNumber}`));
         } else {
-            return this._handleFetchRemoteRun(() => getRemoteData(`/api/runs/${this._runId}`));
+            return this._handleFetchRemoteRun(() => getRemoteData(`/api/legacy/runs/${this._runId}`));
         }
     }
 
@@ -164,7 +164,7 @@ export class RunDetailsModel extends Observable {
             body: JSON.stringify(this.runPatch.toPojo()),
         };
 
-        await this._handleFetchRemoteRun(() => jsonFetch(`/api/v2/runs/${runNumber}`, options));
+        await this._handleFetchRemoteRun(() => jsonFetch(`/api/runs/${runNumber}`, options));
     }
 
     /**
@@ -382,7 +382,7 @@ class RunDetailsTabbedPanelModel extends TabbedPanelModel {
 
         if (this._runNumber) {
             try {
-                const { data: logsOfRun } = await getRemoteData(`/api/v2/runs/${this._runNumber}/logs`);
+                const { data: logsOfRun } = await getRemoteData(`/api/runs/${this._runNumber}/logs`);
                 this.currentPanelData = RemoteData.success(logsOfRun);
             } catch (error) {
                 this.currentPanelData = RemoteData.failure(error);
@@ -402,7 +402,7 @@ class RunDetailsTabbedPanelModel extends TabbedPanelModel {
 
         if (this._runNumber) {
             try {
-                const { data: flpsOfRun } = await getRemoteData(`/api/v2/runs/${this._runNumber}/flps`);
+                const { data: flpsOfRun } = await getRemoteData(`/api/runs/${this._runNumber}/flps`);
                 this.currentPanelData = RemoteData.success(flpsOfRun);
             } catch (error) {
                 this.currentPanelData = RemoteData.failure(error);
