+++ lib/server/routers/runs.router.js
@@ -29,7 +29,7 @@ module.exports = {
                     controller: RunsController.listRuns,
                 },
                 {
-                    path: ':runId',
+                    path: ':runNumber',
                     children: [
                         {
                             method: 'get',
@@ -65,29 +65,15 @@ module.exports = {
                 },
             ],
         }, {
-            path: '/v2/runs',
+            path: '/legacy/runs',
             args: { public: true },
             children: [
                 {
-                    path: ':runNumber',
+                    path: ':runId',
                     children: [
                         {
                             method: 'get',
                             controller: RunsController.getRunByIdentifier,
-                        },
-                        {
-                            method: 'get',
-                            path: 'logs',
-                            controller: RunsController.getLogsByRunIdentifier,
-                        },
-                        {
-                            method: 'get',
-                            path: 'flps',
-                            controller: RunsController.getFlpsByRunIdentifier,
-                        },
-                        {
-                            method: 'put',
-                            controller: RunsController.updateRun,
                         },
                     ],
                 },
