+++ lib/usecases/run/EndRunUseCase.js
@@ -24,15 +24,13 @@ class EndRunUseCase {
      * @returns {Promise<{result: (Run|undefined), error: (Object|undefined)}>} Promise object represents the result of this use case.
      */
     async execute(dto) {
-        const { body, params } = dto;
+        const { body: runPatch, params: { runNumber } } = dto;
 
-        const { lhcPeriod } = body;
-        delete body.lhcPeriod;
+        const { lhcPeriod } = runPatch;
+        delete runPatch.lhcPeriod;
 
-        // Bad variable name, the parameter expect a run number
-        const { runId: runNumber } = params;
         try {
-            return { result: await runService.update({ runNumber }, { runPatch: body, relations: { lhcPeriodName: lhcPeriod } }) };
+            return { result: await runService.update({ runNumber }, { runPatch, relations: { lhcPeriodName: lhcPeriod } }) };
         } catch (e) {
             return {
                 error: {
